# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mural Marketplace API",
    "version": "1.2.0",
    "description": "Backend service powering a USDC marketplace on Polygon via Mural Pay sandbox"
  },
  "servers": [
    { "url": "https://mural-mkt-api.vercel.app", "description": "Production (Vercel)" },
    { "url": "http://localhost:3003", "description": "Local" }
  ],
  "paths": {
...
    "/payouts": {
      "get": {
        "summary": "List all COP withdrawals",
        "description": "Proxies to Mural API. Shows all payout requests and their statuses.",
        "responses": {
          "200": { "description": "Payout list" },
          "502": { "description": "Mural API error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/payouts/{id}": {
      "get": {
        "summary": "Get COP withdrawal status",
        "description": "Proxies to Mural API. Returns live payout status (AWAITING_EXECUTION → PENDING → EXECUTED | FAILED).",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Payout details" },
          "502": { "description": "Mural API error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  }
}